pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = ""
authors = ["Alexander & Ludmila Markiza <alarriva@yandex.ru>"]
readme = "README.md"
packages = [{include = "page_analyzer"}]

[tool.poetry.dependencies]
python = "^3.10"
flask = "^2.3.3"
gunicorn = "^21.2.0"
psycopg2-binary = "^2.9.7"
validators = "^0.22.0"

[tool.poetry.group.dev.dependencies]
flake8 = "^6.1.0"
python-dotenv = "^1.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
